Adds the joinedload() options
This optimizes the number of queries, as I initially fetch all the
tables I know I am going to need.

Note that the query for threads in the view 'vote' turns crazy, this is
because of the LIMIT clause that targets the number of threads and not
the number of rows in the rendered query.